test_task/requirements.txt
Django==5.0.x
djangorestframework==3.15
djangorestframework-simplejwt==5.2.2